requirements.txt
altair==5.5.0
amqp==5.3.1
anyio==3.7.1
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.14.2
billiard==4.2.2
blinker==1.9.0
cachetools==6.2.1
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.2
daphne==4.2.1
deep-translator==1.11.4
dj-database-url==3.0.1
Django==5.2.5
django-allauth==65.13.0
django-appconf==1.2.0
django-celery-beat==2.8.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-image-cropping==1.7
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
fsspec==2025.9.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
gTTS==2.5.4
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.1.1
narwhals==2.12.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
prompt_toolkit==3.0.52
protobuf==6.33.1
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydeck==0.9.1
PyJWT==2.10.1
PyMuPDF==1.26.4
pyOpenSSL==25.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
qrcode==8.2
redis==6.4.0
referencing==0.37.0
reportlab==4.4.4
requests==2.32.5
resend==2.19.0
rpds-py==0.29.0
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
sib-api-v3-sdk==7.6.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
watchdog==6.0.0
watchgod==0.8.2
wcwidth==0.2.14
whitenoise==6.11.0
zope.interface==8.0.1
google-generativeai
streamlit
python-dotenv
